+++ e2e/shared/contracts/cardano/hololocker-demo/aiken.lock
@@ -1,0 +1,15 @@
+# This file was generated by Aiken
+# You typically do not need to edit this file
+
+[[requirements]]
+name = "aiken-lang/stdlib"
+version = "1.6.0"
+source = "github"
+
+[[packages]]
+name = "aiken-lang/stdlib"
+version = "1.6.0"
+requirements = []
+source = "github"
+
+[etags]
