+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "effectstream",
-  "version": "0.100.12",
+  "version": "0.100.13",
   "private": true,
   "workspaces": [
     "packages/**",
